image-agent/requirements.txt
redis==5.2.1
requests==2.32.3
Pillow==11.1.0
boto3==1.37.38
